+++ R/labels_to_levels.R
@@ -79,7 +79,7 @@ labels_to_levels.data.frame <- function(x,
   # create the new variables and updates "select", so new variables are processed
   if (!isFALSE(append)) {
     # process arguments
-    args <- .process_append(
+    arguments <- .process_append(
       x,
       select,
       append,
@@ -89,8 +89,8 @@ labels_to_levels.data.frame <- function(x,
       keep_character = FALSE
     )
     # update processed arguments
-    x <- args$x
-    select <- args$select
+    x <- arguments$x
+    select <- arguments$select
   }
 
   x[select] <- lapply(