+++ NEWS.md
@@ -1,9 +1,32 @@
-# datawizard 0.9.0.9000 (development version)
+# datawizard 0.9.2
+
+CHANGES
+
+* `data_modify()` gets three new arguments, `.at`, `.if` and `.modify`, to modify
+  variables at specific positions or based on logical conditions.
+
+* `data_tabulate()` was revised and gets several new arguments: a `weights`
+  argument, to compute weighted frequency tables. `include_na` allows to include
+  or omit missing values from the table. Furthermore, a `by` argument was added,
+  to compute crosstables (#479, #481).
+
+# datawizard 0.9.1
 
 CHANGES
 
 * `rescale()` gains `multiply` and `add` arguments, to expand ranges by a given
   factor or value.
+
+* `to_factor()` and `to_numeric()` now support class `haven_labelled`.
+
+BUG FIXES
+
+* `to_numeric()` now correctly deals with inversed factor levels when
+  `preserve_levels = TRUE`.
+
+* `to_numeric()` inversed order of value labels when `dummy_factors = FALSE`.
+
+* `convert_to_na()` now preserves attributes for factors when `drop_levels = TRUE`.
 
 # datawizard 0.9.0